Simplify interner type signature w/ associated type
- rm loose interner in favor of using KeyToStr directly
- parsing for multiline branchpoints
- panic when accessing undefined config variable

+++ workflow/src/id.rs
@@ -39,9 +39,11 @@ macro_rules! id {
     };
 }
 
+// TODO if we make sizes variable, these ids need to have type params...
+
 id!(ModuleId, u8);
 id!(BranchpointId, u8);
-id!(IdentId, u8);
+id!(IdentId, u16);
 id!(LiteralId, u8);
 id!(AbstractTaskId, u8);
 id!(AbstractValueId, u8);
